requirements.txt
# This file is provided for convenience. For full dependency management, use:
#   uv sync
#
# See pyproject.toml for the authoritative dependency list.

numpy>=1.26
pandas>=2.2
scikit-learn>=1.5
lightgbm>=4.0
xgboost>=2.0
shap>=0.45
scipy>=1.11
matplotlib>=3.10
rdkit>=2025.3
pydantic>=2.0
python-dotenv>=1.0
httpx>=0.27
anthropic>=0.40
streamlit>=1.38
plotly>=5.22
python-docx>=1.2.0
